feat: move Email class to file and refactor email sender
